requirements.txt
Django==1.8
django-rest-swagger==0.3.5
djangorestframework==3.3.2
gunicorn==19.6.0
